Сохранение опросников в директорию, распарсивание

+++ src/main/java/pharma/lms/PharmaLMS/quiz/repo/QuestionRepo.java
@@ -3,5 +3,5 @@ package pharma.lms.PharmaLMS.quiz.repo;
 import org.springframework.data.jpa.repository.JpaRepository;
 import pharma.lms.PharmaLMS.quiz.domain.Question;
 
-public interface QuestionRepo extends JpaRepository<Question, Long> {
+public interface QuestionRepo extends JpaRepository<Question, Integer> {
 }
